Image Classifier C API: Moved classification results to c/task/processor
